backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aadarsha <aadarshakarki000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.3"
djangorestframework = "^3.16.0"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
celery = "^5.5.3"
gevent = "^25.5.1"
redis = "^6.2.0"
flower = "^2.0.1"
channels = {extras = ["daphne"], version = "^4.2.2"}
django-storages = {extras = ["s3"], version = "^1.14.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
